Add `OpenFunc()` function
This commit adds the `OpenFunc()` function that allows us to handle how
the DB connection will be opened. You can pass any function that has the
same signature as `sql.Open()` and it'll use it to start the connection.

The idea is to open the connections with DataDog [SQL wrapper][1] to
start collecting APM metrics for our queries.

[1]: https://godoc.org/gopkg.in/DataDog/dd-trace-go.v1/contrib/database/sql